+++ libraries/DotMatrix/DotMatrix.h
@@ -17,14 +17,14 @@ class DotMatrix
 {
 public:
     DotMatrix();
-    void setup(int dta, int clk, int cs, int num);
+    void setup(uint8_t dta, uint8_t clk, uint8_t cs, uint8_t num);
     void displayTime(uint8_t hour, uint8_t minute);
     void displayAlarm(bool status);
 private:
     void setColumn(uint8_t col, byte value);
     LedControl lc;
-    uint8_t           last_hour = 100;
-    uint8_t           last_minute = 100;
+    int8_t            last_hour = 100;
+    int8_t            last_minute = 100;
     static byte       const n[10][5];
     static byte       const clock[9];
     static byte       const onoff[2][15];
